+++ internal/css_parser/css_parser.go
@@ -506,7 +506,6 @@ func (p *parser) parseListOfDeclarations(opts listOfDeclarationsOpts) (list []cs
 		case css_lexer.TAtKeyword:
 			if p.inSelectorSubtree > 0 {
 				p.shouldLowerNesting = true
-				p.reportUseOfNesting(p.current().Range, false)
 			}
 			list = append(list, p.parseAtRule(atRuleContext{
 				isDeclarationList:    true,
@@ -525,7 +524,6 @@ func (p *parser) parseListOfDeclarations(opts listOfDeclarationsOpts) (list []cs
 			css_lexer.TDelimGreaterThan,
 			css_lexer.TDelimTilde:
 			p.shouldLowerNesting = true
-			p.reportUseOfNesting(p.current().Range, false)
 			list = append(list, p.parseSelectorRuleFrom(p.index, false, parseSelectorOpts{isDeclarationContext: true}))
 			foundNesting = true
 
@@ -1473,16 +1471,6 @@ func (p *parser) expectValidLayerNameIdent() (string, bool) {
 		return "", false
 	}
 	return text, true
-}
-
-func (p *parser) reportUseOfNesting(r logger.Range, didWarnAlready bool) {
-	if p.options.unsupportedCSSFeatures.Has(compat.Nesting) && p.options.unsupportedCSSFeatures.Has(compat.IsPseudoClass) && !didWarnAlready {
-		text := "CSS nesting syntax is not supported in the configured target environment"
-		if p.options.originalTargetEnv != "" {
-			text = fmt.Sprintf("%s (%s)", text, p.options.originalTargetEnv)
-		}
-		p.log.AddID(logger.MsgID_CSS_UnsupportedCSSNesting, logger.Warning, &p.tracker, r, text)
-	}
 }
 
 func (p *parser) convertTokens(tokens []css_lexer.Token) []css_ast.Token {